Fix stuff from merging development
